Move the version check to FluxInterface base class.
