+++ maestrowf/interfaces/script/_flux/flux0_26_0.py
@@ -22,39 +22,6 @@ class FluxInterface_0260(FluxInterface):
     flux_handle = None
 
     @classmethod
-    def connect_to_flux(cls):
-        if not cls.flux_handle:
-            cls.flux_handle = flux.Flux()
-            LOGGER.debug("New Flux handle created.")
-            broker_version = cls.flux_handle.attr_get("version")
-            adaptor_version = cls.key
-            LOGGER.debug(
-                "Connected to Flux broker running version %s using Maestro "
-                "adapter version %s.", broker_version, adaptor_version)
-            try:
-                from distutils.version import StrictVersion
-                adaptor_version = StrictVersion(adaptor_version)
-                broker_version = StrictVersion(broker_version)
-                if adaptor_version > broker_version:
-                    LOGGER.error(
-                        "Maestro adapter version (%s) is too new for the Flux "
-                        "broker version (%s). Functionality not present in "
-                        "this Flux version may be required by the adapter and "
-                        "cause errors. Please switch to an older adapter.",
-                        adaptor_version, broker_version
-                    )
-                elif adaptor_version < broker_version:
-                    LOGGER.debug(
-                        "Maestro adaptor version (%s) is older than the Flux "
-                        "broker version (%s). This is usually OK, but if a "
-                        "newer Maestro adapter is available, please consider "
-                        "upgrading to maximize performance and compatibility.",
-                        adaptor_version, broker_version
-                    )
-            except ImportError:
-                pass
-
-    @classmethod
     def submit(
         cls, nodes, procs, cores_per_task, path, cwd, walltime,
         ngpus=0, job_name=None, force_broker=True
